chore: move all existing adapters out of views folder
